ess update

+++ src/reservation/reservation.service.ts
@@ -4,47 +4,74 @@ import { Repository } from 'typeorm';
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
-import { ShowService } from 'src/show/show.service';
-import { UserService } from 'src/user/user.service';
 import { Reservation } from './entities/reservation.entity';
+import { User } from 'src/user/entities/user.entity';
+import { Seat } from 'src/show/entities/seat.entity';
+import { Schedule } from 'src/show/entities/schedule.entity';
+import { Show } from 'src/show/entities/show.entity';
 
 @Injectable()
 export class ReservationService {
     constructor(
-        @InjectRepository(Reservation)
-        private reservationRepository: Repository<Reservation>,
-        private readonly showService: ShowService,
-        private readonly userService: UserService,
+        @InjectRepository(Reservation) private readonly reservationRepository: Repository<Reservation>,
+        @InjectRepository(User) private readonly userRepository: Repository<User>,
+        @InjectRepository(Seat) private readonly seatRepository: Repository<Seat>,
+        @InjectRepository(Schedule) private readonly scheduleRepository: Repository<Schedule>,
+        @InjectRepository(Show) private readonly showRepository: Repository<Show>,
     ) { }
 
     //예약 확인하기
-    async getReservation(reservationId: number) {
+    async getReservation(userId: number) {
         return await this.reservationRepository.find({
             order: { createdAt: 'ASC' },
-            where: { reservationId },
+            where: { userId },
         })
     }
 
     //예약하기
-    async reserve(id: number, showId: number, seat: number) {
-        const user = await this.userService.findById(id);
-        const show = await this.showService.findOne(showId);
-        const reservation = await this.reservationRepository.find({ where: { showId } });
+    async reserve(userId: number, scheduleId: number, seat: number) {
+        const user = await this.userRepository.findOneBy({ userId });
+        const schedule = await this.scheduleRepository.findOne({
+            where: { scheduleId },
+        });
+        const seats = await this.seatRepository.findOneBy({ scheduleId });
 
-        if (show.total_seat < show.current_seat) {
-            throw new BadRequestException('남은 좌석이 없습니다.');
+        const show = await this.showRepository.findOneBy({ showId: schedule.showId });
+
+        const price = show.price * seat;
+
+        if (seat > 5) {
+            throw new BadRequestException('1인당 5자리까지 예약할 수 있습니다.');
         }
 
-        if (user.point < show.price) {
+        if (!schedule) {
+            throw new NotFoundException('공연 정보가 없습니다.');
+        }
+
+        if (user.point < price) {
             throw new BadRequestException('보유한 포인트가 부족합니다.')
         }
 
-        if (!reservation) {
-            throw new NotFoundException('해당 공연이 없습니다.');
+        if (seats.current_seat < seat) {
+            throw new BadRequestException('남은 좌석이 부족하거나 없습니다.');
         }
-        await this.reservationRepository.save({ id, showId, seat });
-        this.showService.updateReservation(showId, seat);
-        this.userService.updateReservation(id, show.price);
+
+        //예매 내역 생성
+        const book = await this.reservationRepository.save({ userId, scheduleId });
+
+        console.log(book);
+
+        //포인트 차감
+        const pUser = await this.userRepository.findOneBy({ userId });
+        pUser.point = pUser.point - price;
+        await this.userRepository.save(pUser);
+
+        //좌석 개수 줄이기
+        const rSeat = await this.seatRepository.findOneBy({ scheduleId });
+        rSeat.current_seat -= seat;
+        await this.seatRepository.save(rSeat);
+
+        return book;
     }
 
     // //예약 수정하기
@@ -54,24 +81,39 @@ export class ReservationService {
     // }
 
     //예약 취소하기(포인트 환불)
-    async deleteReservation(reservationId: number, id: number) {
-        await this.verifyMessage(reservationId, id);
-        await this.reservationRepository.delete({ reservationId });
-    }
+    async deleteReservation(reservationId: number, userId: number) {
+        const reservation = await this.reservationRepository.findOneBy({ reservationId });
 
-    private async verifyMessage(reservationId: number, id: number) {
-        const reservation = await this.reservationRepository.findOneBy({
-            reservationId,
-        });
+        const scheduleId = reservation.scheduleId;
 
-        const show = await this.showService.findOne(reservation.showId);
+        const schedule = await this.scheduleRepository.findOneBy({ scheduleId });
 
-        if (_.isNil(reservation) || reservation.reservationId !== id) {
-            throw new NotFoundException(
-                '메시지를 찾을 수 없거나 삭제할 권한이 없습니다.',
-            );
+
+        const showId = schedule.showId;
+
+        console.log("showId");
+        console.log(showId);
+
+        const show = await this.showRepository.findOneBy({ showId });
+
+        const user = await this.userRepository.findOneBy({ userId });
+
+        console.log(show);
+        console.log(user);
+
+        if (_.isNil(reservation)) {
+            throw new NotFoundException('해당 예약 내역이 없습니다.')
         }
 
-        await this.userService.deletePoint(id, show.price);
+        if (user.userId !== userId) {
+            throw new NotFoundException('예약을 취소할 권한이 없습니다.');
+        }
+
+        //예약 취소하기
+        await this.reservationRepository.delete({ reservationId });
+        //포인트 환불
+        await this.userRepository.save({ userId, point: user.point + show.price })
+
+        return show;
     }
 }
